requirements.txt
numpy
discord.py
